+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # Changelog
 
+### 0.2.2
+- Add `get_list_permissions` function
+- Don't raise an exception in the paginator if there is no permission manager
+- Fix context in `get_permission_manager`
+- Fix imports
+
+### 0.2.1
+- Add docstrings
+- Downgrade rest framework requirement to ^3.14
+
 ### 0.2.0
 - Bump permission_manager version.
 - Add documentation.
-